LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_timing_store_and_clear_during_callback-expected.txt


FAIL No entry should be stored in resource timing buffer since its cleared once an item arrived. assert_equals: No entry should be stored in resource timing buffer since its cleared once an item arrived. expected 0 but got 1
PASS 6 resource timing entries should be moved to global buffer. 
FAIL http://localhost:8800/resource-timing/resources/empty.js is not expected to be in the Resource Timing buffer assert_unreached: Reached unreachable code
PASS http://localhost:8800/resource-timing/resources/empty_script.js is expected to have initiatorType script 
PASS http://localhost:8800/resource-timing/resources/webperftestharness.js is expected to have initiatorType script 
PASS http://localhost:8800/resource-timing/resources/webperftestharnessextension.js is expected to have initiatorType script 
PASS http://localhost:8800/resources/testharness.js is expected to have initiatorType script 
PASS http://localhost:8800/resources/testharnessreport.js is expected to have initiatorType script 
FAIL http://localhost:8800/resource-timing/resources/resource_timing_test0.js is expected to be in the Resource Timing buffer assert_unreached: Reached unreachable code
